### Use db migration instead of sync
- [x] create migrations for labels
- [x] create migration for places
- [x] remove sync from db.
- [x] validate the db is in line with the code
- [x] test if all place API calls still work.
