+++ api/package.json
@@ -22,6 +22,7 @@
     "express": "4.17.1",
     "morgan": "1.10.0",
     "sequelize": "6.3.5",
+    "sequelize-cli": "^6.2.0",
     "sqlite3": "5.0.1",
     "winston": "3.3.3",
     "yaml": "1.10.0"
