+++ api/src/models/index.js
@@ -1,12 +1,11 @@
 const Sequelize = require('sequelize');
 const logger = require('../utils/logger');
 
-const Place = require('./place');
-const Label = require('./label');
+const Label = require('./labels');
+const Place = require('./places');
 
 const env = process.env.NODE_ENV || 'development';
-// eslint-disable-next-line import/no-dynamic-require
-const config = require('../../../config/db.json')[env];
+const config = require('../config/config.json')[env];
 
 const db = {};
 
@@ -22,9 +21,27 @@ if (config.use_env_variable) {
 Place.init(sequelize);
 Label.init(sequelize);
 
-sequelize.sync({ alter: true });
-
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
+
+async function isDbUptodate(wanted) {
+  const latest = await sequelize.query('SELECT * FROM sequelizemeta ORDER BY name DESC LIMIT 1;', {
+    type: sequelize.QueryTypes.SELECT,
+  });
+
+  const lName = latest[0].name;
+  const timestamp = lName.substring(0, lName.indexOf('-'));
+
+  /*
+   * At this moment there are no requests being handled so we can exit
+   */
+  if (timestamp !== wanted) {
+    logger.error(`FATAL: database version (${timestamp}) not matching with expected database version(${wanted})!`);
+    logger.info('Did you run the migrate script? See install documentation');
+    process.exit(1);
+  }
+}
+
+isDbUptodate('20210218083610');
 
 module.exports = db;
