+++ api/src/models/labels.js
@@ -1,0 +1,19 @@
+const { DataTypes, Model } = require('sequelize');
+
+class Label extends Model {
+  static init(sequelize) {
+    super.init({
+      state: {
+        type: DataTypes.ENUM,
+        values: ['Locked', 'Assigned', 'Free'],
+        defaultValue: 'Locked',
+        allowNull: false,
+      },
+    }, {
+      sequelize,
+      modelName: 'Labels',
+    });
+  }
+}
+
+module.exports = Label;
